LIME-93: * Extract Theme to enum
