+++ frontend/src/bundles/common/enums/theme.ts
@@ -1,0 +1,6 @@
+const Theme = {
+    DARK: 'dark',
+    LIGHT: 'light',
+} as const;
+
+export { Theme };
