--- frontend/src/bundles/common/types/theme.ts
@@ -1,3 +1,0 @@
-type Theme = 'dark' | 'light';
-
-export { type Theme };
