+++ frontend/src/bundles/common/types/types.ts
@@ -1,6 +1,7 @@
+export { type Theme } from '../enums/theme.js';
 export { type AsyncThunkConfig } from './async-thunk-config.type.js';
 export { type RouteObject } from './route-object.js';
-export { type Theme } from './theme.js';
+export { type PayloadAction } from '@reduxjs/toolkit';
 export { type RouteObject as LibraryRouteObject } from 'react-router-dom';
 export {
     type AuthResponseDto,
